copy clone lint and golint fixed in /cmd - one try
