+++ cmd/interlink/main.go
@@ -2,8 +2,6 @@ package main
 
 import (
 	"context"
-	"crypto/tls"
-	"crypto/x509"
 	"flag"
 	"fmt"
 	"net"
@@ -14,140 +12,17 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/google/uuid"
 	"github.com/sirupsen/logrus"
 	"github.com/virtual-kubelet/virtual-kubelet/log"
 	logruslogger "github.com/virtual-kubelet/virtual-kubelet/log/logrus"
 	"github.com/virtual-kubelet/virtual-kubelet/trace"
 	"github.com/virtual-kubelet/virtual-kubelet/trace/opentelemetry"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/connectivity"
-	"google.golang.org/grpc/credentials"
-	"google.golang.org/grpc/credentials/insecure"
 
+	"github.com/intertwin-eu/interlink/pkg/interlink"
 	types "github.com/intertwin-eu/interlink/pkg/interlink"
 	"github.com/intertwin-eu/interlink/pkg/interlink/api"
 	"github.com/intertwin-eu/interlink/pkg/virtualkubelet"
-
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
-	"go.opentelemetry.io/otel/propagation"
-	"go.opentelemetry.io/otel/sdk/resource"
-	sdktrace "go.opentelemetry.io/otel/sdk/trace"
-	semconv "go.opentelemetry.io/otel/semconv/v1.21.0"
 )
-
-func initProvider(ctx context.Context) (func(context.Context) error, error) {
-	log.G(ctx).Info("Tracing is enabled, setting up the TracerProvider")
-
-	// Get the TELEMETRY_UNIQUE_ID from the environment, if it is not set, use the hostname
-	uniqueID := os.Getenv("TELEMETRY_UNIQUE_ID")
-	if uniqueID == "" {
-		log.G(ctx).Info("No TELEMETRY_UNIQUE_ID set, generating a new one")
-		newUUID := uuid.New()
-		uniqueID = newUUID.String()
-		log.G(ctx).Info("Generated unique ID: ", uniqueID, " use InterLink-Plugin-"+uniqueID+" as service name from Grafana")
-	}
-
-	serviceName := "InterLink-Plugin-" + uniqueID
-
-	res, err := resource.New(ctx,
-		resource.WithAttributes(
-			// the service name used to display traces in backends
-			semconv.ServiceName(serviceName),
-		),
-	)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create resource: %w", err)
-	}
-
-	ctx, cancel := context.WithTimeout(ctx, time.Second)
-	defer cancel()
-
-	otlpEndpoint := os.Getenv("TELEMETRY_ENDPOINT")
-
-	if otlpEndpoint == "" {
-		otlpEndpoint = "localhost:4317"
-	}
-
-	log.G(ctx).Info("TELEMETRY_ENDPOINT: ", otlpEndpoint)
-
-	caCrtFilePath := os.Getenv("TELEMETRY_CA_CRT_FILEPATH")
-
-	conn := &grpc.ClientConn{}
-	if caCrtFilePath != "" {
-
-		// if the CA certificate is provided, set up mutual TLS
-
-		log.G(ctx).Info("CA certificate provided, setting up mutual TLS")
-
-		caCert, err := os.ReadFile(caCrtFilePath)
-		if err != nil {
-			return nil, fmt.Errorf("failed to load CA certificate: %w", err)
-		}
-
-		clientKeyFilePath := os.Getenv("TELEMETRY_CLIENT_KEY_FILEPATH")
-		if clientKeyFilePath == "" {
-			return nil, fmt.Errorf("client key file path not provided. Since a CA certificate is provided, a client key is required for mutual TLS")
-		}
-
-		clientCrtFilePath := os.Getenv("TELEMETRY_CLIENT_CRT_FILEPATH")
-		if clientCrtFilePath == "" {
-			return nil, fmt.Errorf("client certificate file path not provided. Since a CA certificate is provided, a client certificate is required for mutual TLS")
-		}
-
-		certPool := x509.NewCertPool()
-		if !certPool.AppendCertsFromPEM(caCert) {
-			return nil, fmt.Errorf("failed to append CA certificate")
-		}
-
-		cert, err := tls.LoadX509KeyPair(clientCrtFilePath, clientKeyFilePath)
-		if err != nil {
-			return nil, fmt.Errorf("failed to load client certificate: %w", err)
-		}
-
-		tlsConfig := &tls.Config{
-			Certificates:       []tls.Certificate{cert},
-			RootCAs:            certPool,
-			MinVersion:         tls.VersionTLS12,
-			InsecureSkipVerify: false,
-		}
-		creds := credentials.NewTLS(tlsConfig)
-		conn, err = grpc.NewClient(otlpEndpoint, grpc.WithTransportCredentials(creds))
-		if err != nil {
-			return nil, fmt.Errorf("failed to create gRPC connection to collector: %w", err)
-		}
-
-	} else {
-		conn, err = grpc.NewClient(otlpEndpoint, grpc.WithTransportCredentials(insecure.NewCredentials()))
-		if err != nil {
-			return nil, fmt.Errorf("failed to create gRPC connection to collector: %w", err)
-		}
-	}
-
-	conn.WaitForStateChange(ctx, connectivity.Ready)
-
-	// Set up a trace exporter
-	traceExporter, err := otlptracegrpc.New(ctx, otlptracegrpc.WithGRPCConn(conn))
-	if err != nil {
-		return nil, fmt.Errorf("failed to create trace exporter: %w", err)
-	}
-
-	// Register the trace exporter with a TracerProvider, using a batch
-	// span processor to aggregate spans before export.
-	bsp := sdktrace.NewBatchSpanProcessor(traceExporter)
-	tracerProvider := sdktrace.NewTracerProvider(
-		sdktrace.WithSampler(sdktrace.AlwaysSample()),
-		sdktrace.WithResource(res),
-		sdktrace.WithSpanProcessor(bsp),
-	)
-	otel.SetTracerProvider(tracerProvider)
-
-	// set global propagator to tracecontext (the default is no-op).
-	otel.SetTextMapPropagator(propagation.TraceContext{})
-
-	return tracerProvider.Shutdown, nil
-}
 
 func main() {
 	printVersion := flag.Bool("version", false, "show version")
@@ -178,7 +53,7 @@ func main() {
 	defer cancel()
 
 	if os.Getenv("ENABLE_TRACING") == "1" {
-		shutdown, err := initProvider(ctx)
+		shutdown, err := interlink.InitTracer(ctx)
 		if err != nil {
 			log.G(ctx).Fatal(err)
 		}
@@ -190,7 +65,6 @@ func main() {
 
 		log.G(ctx).Info("Tracer setup succeeded")
 
-		// TODO: disable this through options
 		trace.T = opentelemetry.Adapter{}
 	}
